ref(threads-interface): Add threads interface V2
